bugfixes and improvements

+++ src/jtl/Connector/Gambio/Mapper/MeasurementUnitI18n.php
@@ -20,40 +20,49 @@ class MeasurementUnitI18n extends BaseMapper
     {
         $id = null;
 
+        $skip = true;
+
         foreach ($data->getI18ns() as $i18n) {
+            $name = $i18n->getName();
             $language_id = $this->locale2id($i18n->getLanguageISO());
             
             $dbResult = $this->db->query('SELECT code FROM languages WHERE languages_id='.$language_id);
 
             if ($dbResult[0]['code'] == $this->shopConfig['settings']['DEFAULT_LANGUAGE']) {
-                $sql = $this->db->query('SELECT quantity_unit_id FROM quantity_unit_description WHERE language_id='.$language_id.' && unit_name="'.$i18n->getName().'"');
+                $sql = $this->db->query('SELECT quantity_unit_id FROM quantity_unit_description WHERE language_id='.$language_id.' && unit_name="'.$name.'"');
                 if (count($sql) > 0) {
                     $id = $sql[0]['quantity_unit_id'];
                 }
             }
+
+            if (!empty($name)) {
+                $skip = false;
+            }
         }
 
-        if (is_null($id)) {
-            $newUnit = new \stdClass();
-            $id = $this->db->insertRow($newUnit, 'quantity_unit');
-        } else {
-            $this->db->query('DELETE FROM quantity_unit_description WHERE quantity_unit_id='.$id);
+        if ($skip === false) {
+            if (is_null($id)) {
+                $newUnit = new \stdClass();
+                $id = $this->db->insertRow($newUnit, 'quantity_unit');
+            } else {
+                $this->db->query('DELETE FROM quantity_unit_description WHERE quantity_unit_id=' . $id);
+            }
+
+            $data->getId()->setEndpoint($id);
+
+            foreach ($data->getI18ns() as $i18n) {
+                $i18n->getMeasurementUnitId()->setEndpoint($id);
+
+                $unit = new \stdClass();
+                $unit->language_id = $this->locale2id($i18n->getLanguageISO());
+                $unit->quantity_unit_id = $id;
+                $unit->unit_name = $i18n->getName();
+
+                $this->db->insertRow($unit, 'quantity_unit_description');
+            }
+
+            return $data->getI18ns();
         }
-
-        $data->getId()->setEndpoint($id);
-
-        foreach ($data->getI18ns() as $i18n) {
-            $i18n->getMeasurementUnitId()->setEndpoint($id);
-
-            $unit = new \stdClass();
-            $unit->language_id = $this->locale2id($i18n->getLanguageISO());
-            $unit->quantity_unit_id = $id;
-            $unit->unit_name = $i18n->getName();
-
-            $this->db->insertRow($unit, 'quantity_unit_description');            
-        }
-
-        return $data->getI18ns();
     }
 
     protected function languageISO($data)
